src/main/webapp/app/entities/grupa-raportow/route/grupa-raportow-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IGrupaRaportow } from '../grupa-raportow.model';
import { GrupaRaportowService } from '../service/grupa-raportow.service';

import { GrupaRaportowRoutingResolveService } from './grupa-raportow-routing-resolve.service';

describe('GrupaRaportow routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: GrupaRaportowRoutingResolveService;
  let service: GrupaRaportowService;
  let resultGrupaRaportow: IGrupaRaportow | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(GrupaRaportowRoutingResolveService);
    service = TestBed.inject(GrupaRaportowService);
    resultGrupaRaportow = undefined;
  });

  describe('resolve', () => {
    it('should return IGrupaRaportow returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultGrupaRaportow = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultGrupaRaportow).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultGrupaRaportow = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultGrupaRaportow).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IGrupaRaportow>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultGrupaRaportow = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultGrupaRaportow).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
